better chances at safe TLB paddr allocation to integers

+++ Source/Project64/N64 System/Mips/Memory Virtual Mem.h
@@ -166,6 +166,6 @@ private:
 	mutable char m_strLabelName[100];
 
 	//BIG look up table to quickly translate the tlb to real mem address
-	DWORD * m_TLB_ReadMap;
-	DWORD * m_TLB_WriteMap;
+	size_t * m_TLB_ReadMap;
+	size_t * m_TLB_WriteMap;
 };
